+++ src/method.ts
@@ -1,4 +1,5 @@
 export type Method = "get" | "post" | "put" | "delete" | "patch";
+export const methods: Method[] = ["get", "post", "put", "delete", "patch"];
 export type AuxMethod = "options";
 
 export type MethodsDefinition<M extends Method> =