+++ tests/unit/client.spec.ts
@@ -1,0 +1,9 @@
+import { routing } from "../../example/routing";
+import { Client } from "../../src";
+
+describe("API Client Generator", () => {
+  test("Should generate a client for example API", () => {
+    const client = new Client(routing);
+    expect(client.print()).toMatchSnapshot();
+  });
+});