+++ CHANGELOG.md
@@ -2,6 +2,63 @@
 
 ## Version 6
 
+### v6.1.0
+
+- Feature #403: API Client Generator.
+  - A new way of informing the frontend about the I/O types of endpoints.
+  - The new approach offers automatic generation of a client based on routing to a typescript file.
+  - The generated client is flexibly configurable on the frontend side using an implementation function that
+    directly makes requests to an endpoint using the libraries and methods of your choice.
+  - The client asserts the type of request parameters and response.
+- Changes since `beta4`:
+  - Add missing headers to example implementation.
+
+```typescript
+// example client-generator.ts
+import fs from "fs";
+import { Client } from "express-zod-api";
+
+fs.writeFileSync("./frontend/client.ts", new Client(routing).print(), "utf-8");
+```
+
+```typescript
+// example frontend using the most simple Implementation based on fetch
+import { ExpressZodAPIClient } from "./client.ts";
+
+const client = new ExpressZodAPIClient(async (method, path, params) => {
+  const searchParams =
+    method === "get" ? `?${new URLSearchParams(params)}` : "";
+  const response = await fetch(`https://example.com${path}${searchParams}`, {
+    method,
+    headers:
+      method === "get" ? undefined : { "Content-Type": "application/json" },
+    body: method === "get" ? undefined : JSON.stringify(params),
+  });
+  return response.json();
+});
+
+client.provide("get", "/v1/user/retrieve", { id: "10" });
+```
+
+### v6.1.0-beta4
+
+- Path params are excluded from `params` after being substituted.
+
+### v6.1.0-beta3
+
+- The client now accepts a function parameter of `Implementation` type.
+  - Its parameter `path` now contains substituted path params.
+
+### v6.1.0-beta2
+
+- Fixing bugs and taking into account path params.
+
+### v6.1.0-beta1
+
+- This is a beta release of a new feature for public testing.
+- Feature #403: API Client Generator.
+  - More details coming soon.
+
 ### v6.0.3
 
 - `zod` version is 3.14.4.